[Windows] Update wrt-packagemgt-windows-tests to meet the testkitlite requirements
