+++ wrt/wrt-packagemgt-windows-tests/pmwindows/Crosswalk_WebApp_Launch.py
@@ -30,16 +30,21 @@
 
 import unittest
 import os
+import time
 import commands
 import comm
-
 
 class TestWebAppLaunch(unittest.TestCase):
 
     def test_launch(self):
         app_name = "testapp"
         pkg_name = "org.xwalk." + app_name
+
+        if not comm.checkInstalled(pkg_name):
+            comm.app_install(app_name, pkg_name, self)
         comm.app_launch(app_name, pkg_name, self)
+        time.sleep(5)
+        comm.app_stop(self)
 
 if __name__ == '__main__':
     unittest.main()
