+++ wrt/wrt-packagemgt-windows-tests/pmwindows/Crosswalk_WebApp_Stop.py
@@ -39,6 +39,9 @@ class TestWebAppStop(unittest.TestCase):
     def test_stop(self):
         app_name = "testapp"
         pkg_name = "org.xwalk." + app_name
+
+        if not comm.checkInstalled(pkg_name):
+            comm.app_install(app_name, pkg_name, self)
         comm.app_launch(app_name, pkg_name, self)
         comm.app_stop(self)
