+++ wrt/wrt-packagemgt-windows-tests/suite.json
@@ -14,9 +14,8 @@
         "*"
       ],
       "copylist": {
-        "PACK-TOOL-ROOT/resources/inst/inst.msi.py": "inst.py",
-        "PACK-TOOL-ROOT/resources/xsl": ".",
-        "pmwindows": "pmwindows",
+        "inst.msi.py": "inst.py",
+        "README.md": "README.md",
         "tests.full.xml": "tests.full.xml",
         "tests.xml": "tests.xml"
       },
@@ -31,11 +30,11 @@
           "copylist": {
             "../icon.png": "icon.ico"
           },
-          "install-path": "pmwindows",
+          "install-path": ".",
           "apk-url-opt": "http://www.baidu.com"
         }
       }
     }
   },
-  "pkg-name": "wrt-packagemgt-tests"
+  "pkg-name": "wrt-packagemgt-windows-tests"
 }
