+++ wrt/wrt-packagemgt-windows-tests/testapp/README.md
@@ -1,0 +1,15 @@
+## Introduction
+
+This directory is for building "testapp.msi" app
+
+
+## Authors:
+
+* Zhu, YongyongX <yongyongx.zhu@intel.com>
+
+
+## LICENSE
+
+Copyright (c) 2016 Intel Corporation.
+Except as noted, this software is licensed under BSD-3-Clause License.
+Please see the COPYING file for the BSD-3-Clause License.
