+++ wrt/wrt-packagemgt-windows-tests/tests.full.xml
@@ -3,29 +3,29 @@
 <test_definition>
   <suite category="Crosswalk WRT" name="wrt-packagemgt-tests">
     <set name="packagemgt" type="script">
-      <testcase component="Crosswalk WRT/Package Management" execution_type="auto" id="Crosswalk_WebApp_Install" priority="P0" purpose="Validate if Crosswalk support web apps could be installed successfully" status="designed" type="Functional">
+      <testcase component="Crosswalk WRT/Package Management" execution_type="auto" id="Crosswalk_WebApp_Install" priority="P0" purpose="Validate if Crosswalk support web apps could be installed successfully" status="approved" type="Functional">
         <description>
-          <test_script_entry>/opt/wrt-pm-windows-tests/pmwindows/Crosswalk_WebApp_Install.py</test_script_entry>
+          <test_script_entry>python "C:\\Program Files\\wrt-packagemgt-windows-tests\\wrt-packagemgt-windows-tests\\opt\\wrt-packagemgt-windows-tests\\pmwindows\\Crosswalk_WebApp_Install.py"</test_script_entry>
         </description>
       </testcase>
-      <testcase component="Crosswalk WRT/Package Management" execution_type="auto" id="Crosswalk_WebApp_Launch" priority="P0" purpose="Validate if Crosswalk support web apps could be launched successfully" status="designed" type="Functional">
+      <testcase component="Crosswalk WRT/Package Management" execution_type="auto" id="Crosswalk_WebApp_Launch" priority="P0" purpose="Validate if Crosswalk support web apps could be launched successfully" status="approved" type="Functional">
         <description>
-          <test_script_entry>/opt/wrt-pm-windows-tests/pmwindows/Crosswalk_WebApp_Launch.py</test_script_entry>
+          <test_script_entry>python "C:\\Program Files\\wrt-packagemgt-windows-tests\\wrt-packagemgt-windows-tests\\opt\\wrt-packagemgt-windows-tests\\pmwindows\\Crosswalk_WebApp_Launch.py"</test_script_entry>
         </description>
       </testcase>
-      <testcase component="Crosswalk WRT/Package Management" execution_type="auto" id="Crosswalk_WebApp_Stop" priority="P0" purpose="Validate if Crosswalk support web apps could be stoped successfully" status="designed" type="Functional">
+      <testcase component="Crosswalk WRT/Package Management" execution_type="auto" id="Crosswalk_WebApp_Stop" priority="P0" purpose="Validate if Crosswalk support web apps could be stoped successfully" status="approved" type="Functional">
         <description>
-          <test_script_entry>/opt/wrt-pm-windows-tests/pmwindows/Crosswalk_WebApp_Stop.py</test_script_entry>
+          <test_script_entry>python "C:\\Program Files\\wrt-packagemgt-windows-tests\\wrt-packagemgt-windows-tests\\opt\\wrt-packagemgt-windows-tests\\pmwindows\\Crosswalk_WebApp_Stop.py"</test_script_entry>
         </description>
       </testcase>
-      <testcase component="Crosswalk WRT/Package Management" execution_type="auto" id="Crosswalk_WebApp_UnInstall" priority="P0" purpose="Validate if Crosswalk support web apps could be uninstalled successfully" status="designed" type="Functional">
+      <testcase component="Crosswalk WRT/Package Management" execution_type="auto" id="Crosswalk_WebApp_UnInstall" priority="P0" purpose="Validate if Crosswalk support web apps could be uninstalled successfully" status="approved" type="Functional">
         <description>
-          <test_script_entry>/opt/wrt-pm-windows-tests/pmwindows/Crosswalk_WebApp_UnInstall.py</test_script_entry>
+          <test_script_entry>python "C:\\Program Files\\wrt-packagemgt-windows-tests\\wrt-packagemgt-windows-tests\\opt\\wrt-packagemgt-windows-tests\\pmwindows\\Crosswalk_WebApp_UnInstall.py"</test_script_entry>
         </description>
       </testcase>
-      <testcase component="Crosswalk WRT/Package Management" execution_type="auto" id="Crosswalk_WebApp_Update" priority="P0" purpose="Validate if Crosswalk support web apps could be updated successfully" status="designed" type="Functional">
+      <testcase component="Crosswalk WRT/Package Management" execution_type="auto" id="Crosswalk_WebApp_Update" priority="P0" purpose="Validate if Crosswalk support web apps could be updated successfully" status="approved" type="Functional">
         <description>
-          <test_script_entry>/opt/wrt-pm-windows-tests/pmwindows/Crosswalk_WebApp_Update.py</test_script_entry>
+          <test_script_entry>python "C:\\Program Files\\wrt-packagemgt-windows-tests\\wrt-packagemgt-windows-tests\\opt\\wrt-packagemgt-windows-tests\\pmwindows\\Crosswalk_WebApp_Update.py"</test_script_entry>
         </description>
       </testcase>
     </set>
