+++ wrt/wrt-packagemgt-windows-tests/tests.xml
@@ -1,31 +1,31 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <?xml-stylesheet type="text/xsl" href="./testcase.xsl"?>
 <test_definition>
-  <suite category="Crosswalk WRT" name="wrt-packagemgt-tests">
+  <suite category="Crosswalk WRT" name="wrt-packagemgt-windows-tests">
     <set name="packagemgt" type="script">
       <testcase component="Crosswalk WRT/Package Management" execution_type="auto" id="Crosswalk_WebApp_Install" purpose="Validate if Crosswalk support web apps could be installed successfully">
         <description>
-          <test_script_entry>/opt/wrt-pm-windows-tests/pmwindows/Crosswalk_WebApp_Install.py</test_script_entry>
+          <test_script_entry>python "C:\\Program Files\\wrt-packagemgt-windows-tests\\wrt-packagemgt-windows-tests\\opt\\wrt-packagemgt-windows-tests\\pmwindows\\Crosswalk_WebApp_Install.py"</test_script_entry>
         </description>
       </testcase>
       <testcase component="Crosswalk WRT/Package Management" execution_type="auto" id="Crosswalk_WebApp_Launch" purpose="Validate if Crosswalk support web apps could be launched successfully">
         <description>
-          <test_script_entry>/opt/wrt-pm-windows-tests/pmwindows/Crosswalk_WebApp_Launch.py</test_script_entry>
+          <test_script_entry>python "C:\\Program Files\\wrt-packagemgt-windows-tests\\wrt-packagemgt-windows-tests\\opt\\wrt-packagemgt-windows-tests\\pmwindows\\Crosswalk_WebApp_Launch.py"</test_script_entry>
         </description>
       </testcase>
       <testcase component="Crosswalk WRT/Package Management" execution_type="auto" id="Crosswalk_WebApp_Stop" purpose="Validate if Crosswalk support web apps could be stoped successfully">
         <description>
-          <test_script_entry>/opt/wrt-pm-windows-tests/pmwindows/Crosswalk_WebApp_Stop.py</test_script_entry>
+          <test_script_entry>python "C:\\Program Files\\wrt-packagemgt-windows-tests\\wrt-packagemgt-windows-tests\\opt\\wrt-packagemgt-windows-tests\\pmwindows\\Crosswalk_WebApp_Stop.py"</test_script_entry>
         </description>
       </testcase>
       <testcase component="Crosswalk WRT/Package Management" execution_type="auto" id="Crosswalk_WebApp_UnInstall" purpose="Validate if Crosswalk support web apps could be uninstalled successfully">
         <description>
-          <test_script_entry>/opt/wrt-pm-windows-tests/pmwindows/Crosswalk_WebApp_UnInstall.py</test_script_entry>
+          <test_script_entry>python "C:\\Program Files\\wrt-packagemgt-windows-tests\\wrt-packagemgt-windows-tests\\opt\\wrt-packagemgt-windows-tests\\pmwindows\\Crosswalk_WebApp_UnInstall.py"</test_script_entry>
         </description>
       </testcase>
       <testcase component="Crosswalk WRT/Package Management" execution_type="auto" id="Crosswalk_WebApp_Update" purpose="Validate if Crosswalk support web apps could be updated successfully">
         <description>
-          <test_script_entry>/opt/wrt-pm-windows-tests/pmwindows/Crosswalk_WebApp_Update.py</test_script_entry>
+          <test_script_entry>python "C:\\Program Files\\wrt-packagemgt-windows-tests\\wrt-packagemgt-windows-tests\\opt\\wrt-packagemgt-windows-tests\\pmwindows\\Crosswalk_WebApp_Update.py"</test_script_entry>
         </description>
       </testcase>
     </set>
